requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langchain==0.2.6
langchain-openai==0.1.13
langchain-community==0.2.6
faiss-cpu==1.8.0
numpy==1.26.4
rank-bm25==0.2.2
pymupdf==1.24.5
python-multipart==0.0.9
pydantic==2.7.4
python-dotenv==1.0.1
